create a directory for pos_file usage
fluentd is often used to gather log files from directories to which it
does not have write access.  when using the 'tail' source type,
fluentd likes to track the file position in a file provided via the
pos_file parameter.

This patch adds /var/cache/fluentd to the package so that there exists
a directory on the filesystem that can be used for this purpose if
necessary.